Enables HTML content in bells and notifications
Allows bells and notifications to render HTML content in addition to plain text.

This change introduces an `html` property to both the `BellsProps` and `NotificationsProps` interfaces. If provided, the HTML content will be rendered instead of the text. This allows for richer formatting and styling within the notifications and bells.

Adds a warning message if neither `text` nor `html` is provided to ensure that either a text or html prop must be provided.